Fixes to allow CI on boilerplates
