+++ packages/aragon-cli/package.json
@@ -116,6 +116,7 @@
     "nyc": "^14.1.1",
     "prettier": "^1.15.3",
     "proxyquire": "^2.1.0",
+    "ps-tree": "^1.2.0",
     "sinon": "^7.3.1"
   },
   "engines": {
