+++ packages/aragon-cli/test/dao/dao-new.test.js
@@ -1,5 +1,5 @@
 import test from 'ava'
-import { runAragonCLI } from '../utils'
+import { runAragonCLI } from '../util'
 
 const daoAddressRegex = /Created DAO: (.*)$/
 const daoIdAndAddressAddressRegex = /Created DAO: (.*) at (.*)$/
@@ -7,7 +7,7 @@ const daoIdAndAddressAddressRegex = /Created DAO: (.*) at (.*)$/
 test('creates a new DAO', async t => {
   t.plan(1)
 
-  const stdout = await runAragonCLI(['dao', 'new'])
+  const { stdout } = await runAragonCLI(['dao', 'new'])
   const daoAddress = stdout.match(daoAddressRegex)[1]
 
   t.assert(/0x[a-fA-F0-9]{40}/.test(daoAddress), 'Invalid DAO address')
@@ -19,7 +19,7 @@ test('assigns an Aragon Id with the "--aragon-id" param', async t => {
   const date = new Date().getTime()
   const id = `newdao${date}`
 
-  const stdout = await runAragonCLI([
+  const { stdout } = await runAragonCLI([
     'dao',
     'new',
     '--debug',
