+++ packages/aragon-cli/test/deploy/deploy.test.js
@@ -3,7 +3,7 @@ import web3EthAbi from 'web3-eth-abi'
 import { expandLink } from '@aragon/toolkit/dist/util'
 //
 import { deployContract, linkLibraries } from '../../src/lib/deploy/deploy'
-import { isAddress, isValidTxHash, getLocalWeb3 } from '../utils'
+import { isAddress, isValidTxHash, getLocalWeb3 } from '../util'
 
 test.beforeEach(async t => {
   t.context = {
