+++ packages/aragon-cli/test/help/help.test.js
@@ -1,10 +1,12 @@
 import test from 'ava'
-import { runAragonCLI, normalizeOutput } from '../utils'
+//
+import { normalizeOutput } from '../../src/util'
+import { runAragonCLI } from '../util'
 
 test('should return the correct help info', async t => {
   t.plan(1)
 
-  let stdout = await runAragonCLI(['--help'])
+  let { stdout } = await runAragonCLI(['--help'])
   stdout = normalizeOutput(stdout)
 
   t.snapshot(stdout)
