+++ packages/aragon-cli/test/version/snapshots/version.test.js.md
@@ -8,4 +8,12 @@ Generated by [AVA](https://ava.li).
 
 > Snapshot 1
 
-    '6.3.3'
+    {
+      cmd: 'node dist/cli.js --version',
+      code: 0,
+      failed: false,
+      killed: false,
+      signal: null,
+      stderr: '',
+      timedOut: false,
+    }
